OSCORE GET works now

+++ os/net/app-layer/oscore/cose.c
@@ -14,7 +14,7 @@ int cose_encrypt0_encode(cose_encrypt0_t *ptr, uint8_t *buffer){
 }
 	
 /*Return status */
-int cose_encrypt0_decode(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size);
+int cose_encrypt0_decode(cose_encrypt0_t *ptr, uint8_t *buffer, int size);
 
 /* Initiate a new COSE Encrypt0 object. */
 void cose_encrypt0_init(cose_encrypt0_t *ptr){
@@ -24,12 +24,12 @@ void cose_encrypt0_init(cose_encrypt0_t *ptr){
 void cose_encrypt0_set_alg(cose_encrypt0_t *ptr, uint8_t alg){
 	ptr->alg = alg;
 }
-void cose_encrypt0_set_ciphertext(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size){
+void cose_encrypt0_set_ciphertext(cose_encrypt0_t *ptr, uint8_t *buffer, int size){
 	ptr->ciphertext = buffer;
 	ptr->ciphertext_len = size;
 }
 	
-void cose_encrypt0_set_plaintext(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size){
+void cose_encrypt0_set_plaintext(cose_encrypt0_t *ptr, uint8_t *buffer, int size){
 	ptr->plaintext = buffer;
 	ptr->plaintext_len = size;
 }
@@ -37,7 +37,7 @@ void cose_encrypt0_set_plaintext(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t 
 /* Return length */
 int cose_encrypt0_get_plaintext(cose_encrypt0_t *ptr, uint8_t **buffer);
 	
-void cose_encrypt0_set_partial_iv(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size){
+void cose_encrypt0_set_partial_iv(cose_encrypt0_t *ptr, uint8_t *buffer, int size){
 	ptr->partial_iv = buffer;
 	ptr->partial_iv_len = size;
 }
@@ -48,7 +48,7 @@ int cose_encrypt0_get_partial_iv(cose_encrypt0_t *ptr, uint8_t **buffer){
 	return ptr->partial_iv_len;
 }
 	
-void cose_encrypt0_set_key_id(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size){
+void cose_encrypt0_set_key_id(cose_encrypt0_t *ptr, uint8_t *buffer, int size){
 	ptr->key_id = buffer;
 	ptr->key_id_len = size;
 }
@@ -60,13 +60,13 @@ int cose_encrypt0_get_key_id(cose_encrypt0_t *ptr, uint8_t **buffer){
 }
 
 
-void cose_encrypt0_set_external_aad(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size){
+void cose_encrypt0_set_external_aad(cose_encrypt0_t *ptr, uint8_t *buffer, int size){
 	ptr->external_aad = buffer;
 	ptr->external_aad_len = size;
 }
 	
 /* Returns 1 if successfull, 0 if key is of incorrect length. */
-int cose_encrypt0_set_key(cose_encrypt0_t *ptr, uint8_t *key, uint8_t key_size){
+int cose_encrypt0_set_key(cose_encrypt0_t *ptr, uint8_t *key, int key_size){
 	if(key_size != 16){
 		return 0;
 	}
@@ -77,13 +77,13 @@ int cose_encrypt0_set_key(cose_encrypt0_t *ptr, uint8_t *key, uint8_t key_size){
 	return 1;
 }
 	
-void cose_encrypt0_set_nonce(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size){
+void cose_encrypt0_set_nonce(cose_encrypt0_t *ptr, uint8_t *buffer, int size){
 	ptr->nonce = buffer;
 	ptr->nonce_len = size;
 }
 
 
-int cose_encrypt0_encrypt(cose_encrypt0_t *ptr, uint8_t *ciphertext_buffer, uint8_t ciphertext_len){
+int cose_encrypt0_encrypt(cose_encrypt0_t *ptr, uint8_t *ciphertext_buffer, int ciphertext_len){
 	if(ptr->key == NULL || ptr->key_len != 16){
 		return -1;
 	}
@@ -102,7 +102,7 @@ int cose_encrypt0_encrypt(cose_encrypt0_t *ptr, uint8_t *ciphertext_buffer, uint
 }
 
 
-int cose_encrypt0_decrypt(cose_encrypt0_t *ptr, uint8_t *plaintext_buffer, uint8_t plaintext_len){
+int cose_encrypt0_decrypt(cose_encrypt0_t *ptr, uint8_t *plaintext_buffer, int plaintext_len){
 	if(ptr->key == NULL || ptr->key_len != 16){
 		return -1;
 	}
