+++ os/net/app-layer/oscore/cose.h
@@ -40,36 +40,36 @@ typedef struct cose_encrypt0_t {
 int cose_encrypt0_encode(cose_encrypt0_t *ptr, uint8_t *buffer);
 	
 /*Return status */
-int cose_encrypt0_decode(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size);
+int cose_encrypt0_decode(cose_encrypt0_t *ptr, uint8_t *buffer, int size);
 
 /* Initiate a new COSE Encrypt0 object. */
 void cose_encrypt0_init(cose_encrypt0_t *ptr);	
 
 void cose_encrypt0_set_alg(cose_encrypt0_t *ptr, uint8_t alg);
 	
-void cose_encrypt0_set_plaintext(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size);
-void cose_encrypt0_set_ciphertext(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size);
+void cose_encrypt0_set_plaintext(cose_encrypt0_t *ptr, uint8_t *buffer, int size);
+void cose_encrypt0_set_ciphertext(cose_encrypt0_t *ptr, uint8_t *buffer, int size);
 /* Return length */
 int cose_encrypt0_get_plaintext(cose_encrypt0_t *ptr, uint8_t **buffer);
 	
-void cose_encrypt0_set_partial_iv(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size);
+void cose_encrypt0_set_partial_iv(cose_encrypt0_t *ptr, uint8_t *buffer, int size);
 	
 /* Return length */
 int cose_encrypt0_get_partial_iv(cose_encrypt0_t *ptr, uint8_t **buffer);
 	
-void cose_encrypt0_set_key_id(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size);
+void cose_encrypt0_set_key_id(cose_encrypt0_t *ptr, uint8_t *buffer, int size);
 	
 /* Return length */	
 int cose_encrypt0_get_key_id(cose_encrypt0_t *ptr, uint8_t **buffer);
 
-void cose_encrypt0_set_external_aad(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size);
+void cose_encrypt0_set_external_aad(cose_encrypt0_t *ptr, uint8_t *buffer, int size);
 	
 /* Returns 1 if successfull, 0 if key is of incorrect length. */
-int cose_encrypt0_set_key(cose_encrypt0_t *ptr, uint8_t *key, uint8_t key_size);
+int cose_encrypt0_set_key(cose_encrypt0_t *ptr, uint8_t *key, int key_size);
 	
-void cose_encrypt0_set_nonce(cose_encrypt0_t *ptr, uint8_t *buffer, uint8_t size);
+void cose_encrypt0_set_nonce(cose_encrypt0_t *ptr, uint8_t *buffer, int size);
 
-int cose_encrypt0_encrypt(cose_encrypt0_t *ptr, uint8_t *ciphertext_buffer, uint8_t ciphertext_len);
-int cose_encrypt0_decrypt(cose_encrypt0_t *ptr, uint8_t *plaintext_buffer, uint8_t plaintext_len);
+int cose_encrypt0_encrypt(cose_encrypt0_t *ptr, uint8_t *ciphertext_buffer, int ciphertext_len);
+int cose_encrypt0_decrypt(cose_encrypt0_t *ptr, uint8_t *plaintext_buffer, int plaintext_len);
 
 #endif /* _COSE_H */
